ast errors fixed and some more fixes😬

+++ driver.c
@@ -134,8 +134,8 @@ int main(int argc, char *argv[])
 					end_time = clock();
 					if(!m1)
 					{ 
-						total_CPU_time +=  (double) (end_time - start_time);
-                		total_CPU_time_in_seconds +=  total_CPU_time / CLOCKS_PER_SEC;
+						total_CPU_time += ((double)end_time - start_time);
+                		total_CPU_time_in_seconds += (total_CPU_time / CLOCKS_PER_SEC);
 						nodeCount(root, &count);
 						count++; 
 					}	
@@ -155,8 +155,8 @@ int main(int argc, char *argv[])
 				end_time = clock();
 				if(!m2)
 					{ 
-						total_CPU_time +=  (double) (end_time - start_time);
-                		total_CPU_time_in_seconds +=  total_CPU_time / CLOCKS_PER_SEC;
+						total_CPU_time += ((double)end_time - start_time);
+                		total_CPU_time_in_seconds += (total_CPU_time / CLOCKS_PER_SEC);
 						nodeCountAST(ast, &countAST);
 						countAST++;
 					}	
@@ -170,8 +170,8 @@ int main(int argc, char *argv[])
 				}
 				printf("\n\nParse Tree number of nodes= %d\tAllocated Memory = %ld Bytes\n",count,count*sizeof(parsetree));
 		     		printf("\nAST number of nodes= %d\tAllocated Memory = %ld Bytes\n",countAST,countAST*sizeof(parsetree));	
-		     		float compressionPerc = (float)((count-countAST)/count)*100 ;
-				printf("Compression Percentage = %f\n\n\n",compressionPerc);
+		     		double compressionPerc = (((double)count-countAST)/count)*100 ;
+				printf("Compression Percentage = %lf\n\n\n",compressionPerc);
 				printf("\n\n------------------------------------------------------------------\n\n");
 		     		break;
 
@@ -188,8 +188,8 @@ int main(int argc, char *argv[])
 		     			start_time = clock();
 		     			st = fillSymbolTable(ast);
 		     			end_time = clock();
-		     			total_CPU_time  +=  (double) (end_time - start_time);
-                			total_CPU_time_in_seconds  +=  total_CPU_time / CLOCKS_PER_SEC;
+		     			total_CPU_time  += ((double)end_time - start_time);
+                        total_CPU_time_in_seconds += (total_CPU_time / CLOCKS_PER_SEC);
 		     		}
 		     		m4++;
 		     		printSymbolTable(st);
@@ -240,20 +240,20 @@ int main(int argc, char *argv[])
 		     		{
 		     			start_time = clock();
 		     			TypeChecker(*ast, st);
-					checkFunctionSemantics(ast, st);
+                        checkFunctionSemantics(ast, st);
 		     			end_time = clock();
-		     			total_CPU_time  +=  (double) (end_time - start_time);
-                			total_CPU_time_in_seconds  +=  total_CPU_time / CLOCKS_PER_SEC;
+		     			total_CPU_time  += ((double)end_time - start_time);
+                			total_CPU_time_in_seconds += (total_CPU_time / CLOCKS_PER_SEC);
 		     		}
 				m5++;
 				printf("\n\ntotal_CPU_time:%lf\n",total_CPU_time);
-                		printf("total_CPU_time_in_seconds:%lf  seconds\n\n",total_CPU_time_in_seconds);
-		     		printf("\n\n------------------------------------------------------------------\n\n");
-		     		break;
+                printf("total_CPU_time_in_seconds:%lf  seconds\n\n",total_CPU_time_in_seconds);
+                printf("\n\n------------------------------------------------------------------\n\n");
+                break;
 		
 		     	default:
 		     		// invalid option
-		            	fprintf(stderr, "\nstage2: error: invalid option\n");
+                    fprintf(stderr, "\nstage2: error: invalid option\n");
         }
     }
